requirements.txt
# AWS dependencies
aws-cdk-lib==2.215.0
constructs>=10.0.0,<11.0.0

# Virtual environment management
python-decouple==3.8
